pyproject.toml
[project]
name = "arxiv-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.18",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.4.7",
    "openai>=1.82.0",
    "pdfplumber>=0.11.6",
    "pymupdf4llm>=0.0.24",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.2",
    "streamlit>=1.45.1",
    "tiktoken>=0.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
